Balancing functions

This file describes the usage of all the functions under the balancing tab in the ENTSOE-E transparancy platform.

User functions

entsoeAPI.current_balancing_state — Function

current_balancing_state(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the current balancing state in a certain area (article GL EB 12.3 A). Parses the received HTTP response and returns the data in a dataframe.

[start, end, situation, open loop ace]

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! 100 day range limit applies !

entsoeAPI.balancing_energy_bids — Function

balancing_energy_bids(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime, processType::String)

Constructs the HTTP request for the data about the balancing energy bids in a certain area (article GL EB 12.3 B-D). Parses the received HTTP response and returns the data in a dataframe.

NOT IMPLEMENTED YET

Arguments

processType::String: identifies the type of processing to be carried out on the information

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! 1 day range limit applies !

entsoeAPI.aggregated_balancing_energy_bids — Function

aggregated_balancing_energy_bids(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime, processType::String)

Constructs the HTTP request for the data about the aggregated balancing energy bids in a certain area (article GL EB 12.3 E: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Aggregated%20Bids.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, type of product, direction, offered, activated, unavailable]

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

processType::String: identifies the type of processing to be carried out on the information

! One year range limit applies !

entsoeAPI.procured_balancing_capacity — Function

procured_balancing_capacity(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, type_MarketAgreementType::String = ""])

Constructs the HTTP request for the data about the produced balancing capacity in a certain area (article GL EB 12.3 F: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Procured%20Capacity.html). Parses the received HTTP response and returns the data in a dataframe.

NOT IMPLEMENTED YET

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

type_MarketAgreementType::String = "": Indicates the time horizon for which balancing capacity was procured

! 100 document limit applies !

entsoeAPI.crossZonal_balancing_capacity — Function

crossZonal_balancing_capacity(acquiring_Domain::Union{mappings.Area, String}, connecting_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the use of the allocated cross-zonal balancing capacity over a certain border (article GL EB 12.3 H&I: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Use%20of%20allocated%20cross-zonal%20balancing%20capacity.html). Parses the received HTTP response and returns the data in a dataframe.

NOT IMPLEMENTED YET

Arguments

acquiring_Domain::Union{mappings.Area, String}:

connecting_Domain::Union{mappings.Area, String}:

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

type_MarketAgreementType::String = "": Indicates the time horizon for which balancing capacity was procured

entsoeAPI.volumes_and_prices_contracted_reserves — Function

volumes_and_prices_contracted_reserves(type_MarketAgreementType::String, processType::String, controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the prices and the volumes of the contracted reserves in a certain area (article 17.1 B&C). Parses the received HTTP response and returns the data in a dataframe.

NOT IMPLEMENTED YET

Minimum time interval in query response ranges from part of day to year, depending on selected Type_MarketAgreement.Type!

Arguments

type_MarketAgreementType::String: Indicates the time horizon for which balancing capacity was procured

processType::String: identifies the type of processing to be carried out on the information

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

offset::Int = 0: allows downloading more than 100 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+100

! 100 document limit applies !

entsoeAPI.volumes_contracted_reserves — Function

volumes_contracted_reserves(type_MarketAgreementType::String, controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the volumes of the contracted reserves in a certain area (article 17.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Volumes%20of%20Contracted%20Balancing%20Reserves.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, reserve type, regulation volume, direction]

Minimum time interval in query response ranges from part of day to year, depending on selected Type_MarketAgreement.Type!

Arguments

type_MarketAgreementType::String: Indicates the time horizon for which balancing capacity was procured

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

offset::Int = 0: allows downloading more than 100 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+100

! 100 document limit applies ! ! Doesn't work for hourly data due to some unclear reason... !

entsoeAPI.prices_contracted_reserves — Function

prices_contracted_reserves(type_MarketAgreementType::String, controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the prices of the contracted reserves in a certain area (article 17.1 C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Price%20of%20Reserved%20Balancing%20Reserves.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, reserve type, regulation price, direction, price type]

Minimum time interval in query response ranges from part of day to year, depending on selected Type_MarketAgreement.Type!

Arguments

type_MarketAgreementType::String: Indicates the time horizon for which balancing capacity was procured

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

offset::Int = 0: allows downloading more than 100 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+100

! 100 document limit applies !

entsoeAPI.accepted_aggregated_offers — Function

accepted_aggregated_offers(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""])

Constructs the HTTP request for the data about the accepted aggregated offers in a certain area (article 17.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Accepted%20Offers%20and%20Activated%20Balancing%20Reserves.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, reserve type, source, activated reserves, direction]

Minimum time interval in query response is one BTU period!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

! One year range limit applies !

entsoeAPI.activated_balancing_energy — Function

activated_balancing_energy(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""])

Constructs the HTTP request for the data about the activated balancing energy in a certain area (article 17.1 E: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Accepted%20Offers%20and%20Activated%20Balancing%20Reserves.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, reserve type, source, activated energy, direction]

Minimum time interval in query response is one BTU period!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

! One year range limit applies !

entsoeAPI.prices_activated_balancing_energy — Function

prices_activated_balancing_energy(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""])

Constructs the HTTP request for the data about the prices of the activated balancing energy in a certain area (article 17.1 F: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Accepted%20Offers%20and%20Activated%20Balancing%20Reserves.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, reserve type, source, price type, price, direction]

Minimum time interval in query response is one BTU period!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

! One year range limit applies !

entsoeAPI.imbalance_prices — Function

imbalance_prices(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the imbalance prices in a certain area (article 17.1 G: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Imbalance.html). Parses the received HTTP response and returns the data in a dataframe.

NOT IMPLEMENTED YET

Minimum time interval in query response is one BTU period!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.total_imbalance_volumes — Function

total_imbalance_volumes(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the imbalance volumes in a certain area (article 17.1 H: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Imbalance.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, volume, difference, situation, status]

Minimum time interval in query response is one BTU period!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.financial_expenses — Function

financial_expenses(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the financial expenses and income for balancing in a certain area (article 17.1 I: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Financial%20Expenses%20and%20Income.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, income, expenses, status]

Minimum time interval in query response is one month!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.crossBorder_balancing — Function

crossBorder_balancing(acquiring_Domain::Union{mappings.Area, String}, connecting_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the cross-border balancing over a certain border (article 17.1 J: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Cross-Border%20Balancing.html). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Minimum time interval in query response is one BTU period!

Arguments

acquiring_Domain::Union{mappings.Area, String}:

connecting_Domain::Union{mappings.Area, String}:

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.FCR_total_capacity — Function

FCR_total_capacity(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the FCR total capacity in a certain area (article SO GL 187.2). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.share_capacity_FCR — Function

share_capacity_FCR(area_Domain::Union{mappings.Area, String}, periodStart::Datetime, periodEnd::DateTime)

Constructs the HTTP request for the data about the share of FCR capacity in a certain area (article SO GL 187.2). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.contracted_reserve_capacity_FCR — Function

contracted_reserve_capacity_FCR(area_Domain::Union{mappings.Area}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the share of contracted reserve FCR capacity in a certain area (article SO GL 187.2). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.FRR_actual_capacity — Function

FRR_actual_capacity(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the actual FRR capacity in a certain area (article SO GL 188.4). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One quarter range limit applies !

entsoeAPI.RR_actual_capacity — Function

RR_actual_capacity(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the actual RR capacity in a certain area (article SO GL 189.3). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One quarter range limit applies !

entsoeAPI.sharing_of_reserves — Function

sharing_of_reserves(processType::String, acquiring_Domain::Union{mappings.Area, String}, connecting_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the sharing of RR and FRR over a certain border (article SO GL 190.1). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Arguments

processType::String: identifies the type of processing to be carried out on the information

acquiring_Domain::Union{mappings.Area, String}:

connecting_Domain::Union{mappings.Area, String}:

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.balancing_border_capacity_limitations — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.permanent_allocation_limitations_HVDC — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.netted_and_exchanged_volumes — Function

STILL NEEDS TO BE IMPLEMENTED

GET functions

entsoeAPI.GETconstructor.query_current_balancing_state — Function

query_current_balancing_state(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the current balancing state in a certain area (article GL EB 12.3 A). Returns the received HTTP response, together with the used timezone.

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! 100 day range limit applies !

entsoeAPI.GETconstructor.query_balancing_energy_bids — Function

query_balancing_energy_bids(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime, processType::String)

Constructs the HTTP request for the data about the balancing energy bids in a certain area (article GL EB 12.3 B-D). Returns the received HTTP response, together with the used timezone.

Arguments

processType::String: identifies the type of processing to be carried out on the information

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! 1 day range limit applies !

entsoeAPI.GETconstructor.query_aggregated_balancing_energy_bids — Function

query_aggregated_balancing_energy_bids(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime, processType::String)

Constructs the HTTP request for the data about the aggregated balancing energy bids in a certain area (article GL EB 12.3 E: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Aggregated%20Bids.html). Returns the received HTTP response, together with the used timezone.

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

processType::String: identifies the type of processing to be carried out on the information

! One year range limit applies !

entsoeAPI.GETconstructor.query_procured_balancing_capacity — Function

query_procured_balancing_capacity(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, type_MarketAgreementType::String = ""])

Constructs the HTTP request for the data about the produced balancing capacity in a certain area (article GL EB 12.3 F: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Procured%20Capacity.html). Returns the received HTTP response, together with the used timezone.

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

type_MarketAgreementType::String = "": Indicates the time horizon for which balancing capacity was procured

! 100 document limit applies !

entsoeAPI.GETconstructor.query_crossZonal_balancing_capacity — Function

query_crossZonal_balancing_capacity(acquiring_Domain::Union{mappings.Area, String}, connecting_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the use of the allocated cross-zonal balancing capacity over a certain border (article GL EB 12.3 H&I: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Use%20of%20allocated%20cross-zonal%20balancing%20capacity.html). Returns the received HTTP response, together with the used timezone.

Arguments

acquiring_Domain::Union{mappings.Area, String}:

connecting_Domain::Union{mappings.Area, String}:

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

type_MarketAgreementType::String = "": Indicates the time horizon for which balancing capacity was procured

entsoeAPI.GETconstructor.query_volumes_and_prices_contracted_reserves — Function

query_volumes_and_prices_contracted_reserves(type_MarketAgreementType::String, processType::String, controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the prices and the volumes of the contracted reserves in a certain area (article 17.1 B&C). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response ranges from part of day to year, depending on selected Type_MarketAgreement.Type!

Arguments

type_MarketAgreementType::String: Indicates the time horizon for which balancing capacity was procured

processType::String: identifies the type of processing to be carried out on the information

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

offset::Int = 0: allows downloading more than 100 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+100

! 100 document limit applies !

entsoeAPI.GETconstructor.query_volumes_contracted_reserves — Function

query_volumes_contracted_reserves(type_MarketAgreementType::String, controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the volumes of the contracted reserves in a certain area (article 17.1 B: https://transparency.entsoe.eu/content/staticcontent/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Volumes%20of%20Contracted%20Balancing%20Reserves.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response ranges from part of day to year, depending on selected TypeMarketAgreement.Type!

Arguments

type_MarketAgreementType::String: Indicates the time horizon for which balancing capacity was procured

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

offset::Int = 0: allows downloading more than 100 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+100

! 100 document limit applies ! ! Doesn't work for hourly data due to some unclear reason... !

entsoeAPI.GETconstructor.query_prices_contracted_reserves — Function

query_prices_contracted_reserves(type_MarketAgreementType::String, controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the prices of the contracted reserves in a certain area (article 17.1 C: https://transparency.entsoe.eu/content/staticcontent/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Price%20of%20Reserved%20Balancing%20Reserves.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response ranges from part of day to year, depending on selected TypeMarketAgreement.Type!

Arguments

type_MarketAgreementType::String: Indicates the time horizon for which balancing capacity was procured

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

offset::Int = 0: allows downloading more than 100 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+100

! 100 document limit applies !

entsoeAPI.GETconstructor.query_accepted_aggregated_offers — Function

query_accepted_aggregated_offers(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""])

Constructs the HTTP request for the data about the accepted aggregated offers in a certain area (article 17.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Accepted%20Offers%20and%20Activated%20Balancing%20Reserves.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one BTU period!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

! One year range limit applies !

entsoeAPI.GETconstructor.query_activated_balancing_energy — Function

query_activated_balancing_energy(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""])

Constructs the HTTP request for the data about the activated balancing energy in a certain area (article 17.1 E: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Accepted%20Offers%20and%20Activated%20Balancing%20Reserves.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one BTU period!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

! One year range limit applies !

entsoeAPI.GETconstructor.query_prices_activated_balancing_energy — Function

query_prices_activated_balancing_energy(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""])

Constructs the HTTP request for the data about the prices of the activated balancing energy in a certain area (article 17.1 F: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Accepted%20Offers%20and%20Activated%20Balancing%20Reserves.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one BTU period!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

! One year range limit applies !

entsoeAPI.GETconstructor.query_imbalance_prices — Function

query_imbalance_prices(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the imbalance prices in a certain area (article 17.1 G: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Imbalance.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one BTU period!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_total_imbalance_volumes — Function

query_total_imbalance_volumes(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the imbalance volumes in a certain area (article 17.1 H: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Imbalance.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one BTU period!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_financial_expenses — Function

query_financial_expenses(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the financial expenses and income for balancing in a certain area (article 17.1 I: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Financial%20Expenses%20and%20Income.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one month!

Arguments

controlArea_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_crossBorder_balancing — Function

query_crossBorder_balancing(acquiring_Domain::Union{mappings.Area, String}, connecting_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the cross-border balancing over a certain border (article 17.1 J: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/balancing/Data-view%20Cross-Border%20Balancing.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one BTU period!

Arguments

acquiring_Domain::Union{mappings.Area, String}:

connecting_Domain::Union{mappings.Area, String}:

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_FCR_total_capacity — Function

query_FCR_total_capacity(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the FCR total capacity in a certain area (article SO GL 187.2). Returns the received HTTP response.

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_share_capacity_FCR — Function

query_share_capacity_FCR(area_Domain::Union{mappings.Area, String}, periodStart::Datetime, periodEnd::DateTime)

Constructs the HTTP request for the data about the share of FCR capacity in a certain area (article SO GL 187.2). Returns the received HTTP response.

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_contracted_reserve_capacity_FCR — Function

query_contracted_reserve_capacity_FCR(area_Domain::Union{mappings.Area}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the share of contracted reserve FCR capacity in a certain area (article SO GL 187.2). Returns the received HTTP response.

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_FRR_actual_capacity — Function

query_FRR_actual_capacity(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the actual FRR capacity in a certain area (article SO GL 188.4). Returns the received HTTP response.

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One quarter range limit applies !

entsoeAPI.GETconstructor.query_RR_actual_capacity — Function

query_RR_actual_capacity(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the actual RR capacity in a certain area (article SO GL 189.3). Returns the received HTTP response.

Arguments

area_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One quarter range limit applies !

entsoeAPI.GETconstructor.query_sharing_of_reserves — Function

query_sharing_of_reserves(processType::String, acquiring_Domain::Union{mappings.Area, String}, connecting_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the sharing of RR and FRR over a certain border (article SO GL 190.1). Returns the received HTTP response.

Arguments

processType::String: identifies the type of processing to be carried out on the information

acquiring_Domain::Union{mappings.Area, String}:

connecting_Domain::Union{mappings.Area, String}:

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_balancing_border_capacity_limitations — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.GETconstructor.query_permanent_allocation_limitations_HVDC — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.GETconstructor.query_netted_and_exchanged_volumes — Function

STILL NEEDS TO BE IMPLEMENTED

Parse functions

entsoeAPI.xmlParser.parse_current_balancing_state — Function

parse_current_balancing_state(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_current_balancing_state(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[start, end, situation, open loop ace]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_current_balancing_state(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_balancing_energy_bids — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_aggregated_balancing_energy_bids — Function

parse_aggregated_balancing_energy_bids(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_aggregated_balancing_energy_bids(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime, processType::String). Returns the data in a dataframe.

[start, end, type of product, direction, offered, activated, unavailable]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_aggregated_balancing_energy_bids(area_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime, processType::String)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_procured_balancing_capacity — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_crossZonal_balancing_capacity — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_volumes_and_prices_contracted_reserves — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_volumes_contracted_reserves — Function

parse_volumes_contracted_reserves(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_volumes_contracted_reserves(type_MarketAgreementType::String, controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = "", offset::Int = 0]). Returns the data in a dataframe.

[start, end, reserve type, source, regulation volume, direction]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_volumes_contracted_reserves(type_MarketAgreementType::String, controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = "", offset::Int = 0])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_prices_contracted_reserves — Function

parse_prices_contracted_reserves(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_prices_contracted_reserves(type_MarketAgreementType::String, controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = "", offset::Int = 0]). Returns the data in a dataframe.

[start, end, reserve type, source, regulation price, direction, price type]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_prices_contracted_reserves(type_MarketAgreementType::String, controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = "", offset::Int = 0])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_accepted_aggregated_offers — Function

parse_accepted_aggregated_offers(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_accepted_aggregated_offers(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""]). Returns the data in a dataframe.

[start, end, reserve type, source, accepted reserves, direction]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_accepted_aggregated_offers(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_activated_balancing_energy — Function

parse_activated_balancing_energy(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_activated_balancing_energy(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""]). Returns the data in a dataframe.

[start, end, reserve type, source, activated energy, direction]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_activated_balancing_energy(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_prices_activated_balancing_energy — Function

parse_prices_activated_balancing_energy(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_prices_activated_balancing_energy(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""]). Returns the data in a dataframe.

[start, end, reserve type, source, price type, price, direction]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_prices_activated_balancing_energy(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", psrType::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_imbalance_prices — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_total_imbalance_volumes — Function

parse_total_imbalance_volumes(xml, tz)

Parses xml file generated by the function GETconstructor.query_total_imbalance_volumes(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[start, end, volume, difference, situation, status]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_total_imbalance_volumes(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_financial_expenses — Function

parse_financial_expenses(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_financial_expenses(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[start, end, income, expenses, status]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_financial_expenses(controlArea_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_crossBorder_balancing — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_FCR_total_capacity — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_share_capacity_FCR — Function

STILL NEEDS TO BE IMPLEMENTED

Missing docstring.

Missing docstring for entsoeAPI.xmlParser.parse_contracted_reserve_capacity_FCR. Check Documenter's build log for details.

entsoeAPI.xmlParser.parse_FRR_actual_capacity — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_RR_actual_capacity — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_sharing_of_reserves — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_balancing_border_capacity_limitations — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_permanent_allocation_limitations_HVDC — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_netted_and_exchanged_volumes — Function

STILL NEEDS TO BE IMPLEMENTED