Congestion management functions

This file describes the usage of all the functions under the congestion management tab in the ENTSOE-E transparancy platform.

User functions

entsoeAPI.expansion_and_dismantling — Function

expansion_and_dismantling(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", docStatus::String = ""])

Constructs the HTTP request for the data about expansion and dismantling projects over a certain border (article 9.1: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission/Data-view%20Expansion%20and%20dismantling%20projects.html). Parses the received HTTP response and returns the data in a dataframe.

[estimated completion date, new NTC => [start, end, new NTC], transmission assets => [code, location, type]]

Time interval in query response depends on duration of matching projects!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

docStatus::String = "": Identification of the condition or position of the document with regard to its standing

! 100 document limit applies !

entsoeAPI.redispatching — Function

redispatching(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = ""])

Constructs the HTTP request for the data about redispatching over a certain border (article 13.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/congestion-management/Data-view%20Redispatching.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, reason, impact => [start, end, impact], affected assets => [code, location, type]]

Time interval in query response depends on duration of matching redispatches!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

! 100 document limit applies !

entsoeAPI.countertrading — Function

countertrading(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about countertrading over a certain border (article 13.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/congestion-management/Data-view%20Countertrading.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, reason, change in cross-border exchange => [start, end, change in cross-border exchange]]

Time interval in query response depends on duration of matching counter trades!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

! 100 document limit applies !

entsoeAPI.congestion_costs — Function

congestion_costs(domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = ""])

Constructs the HTTP request for the data about the congestion management costs over a certain border (article 13.1 C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/congestion-management/Data-view%20Costs.html). Parses the received HTTP response and returns the data in a dataframe.

[start, end, costs]

Minimum time interval in query response is one month!

Arguments

domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

! 100 document limit applies !

GET functions

entsoeAPI.GETconstructor.query_expansion_and_dismantling — Function

query_expansion_and_dismantling(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", docStatus::String = ""])

Constructs the HTTP request for the data about expansion and dismantling projects over a certain border (article 9.1: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission/Data-view%20Expansion%20and%20dismantling%20projects.html). Returns the received HTTP response, together with the used timezone. Time interval in query response depends on duration of matching projects!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

docStatus::String = "": Identification of the condition or position of the document with regard to its standing

! 100 document limit applies !

entsoeAPI.GETconstructor.query_redispatching — Function

query_redispatching(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = ""])

Constructs the HTTP request for the data about redispatching over a certain border (article 13.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/congestion-management/Data-view%20Redispatching.html). Returns the received HTTP response, together with the used timezone. Time interval in query response depends on duration of matching redispatches!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

! 100 document limit applies !

entsoeAPI.GETconstructor.query_countertrading — Function

query_countertrading(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about countertrading over a certain border (article 13.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/congestion-management/Data-view%20Countertrading.html). Returns the received HTTP response, together with the used timezone. Time interval in query response depends on duration of matching counter trades!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

! 100 document limit applies !

entsoeAPI.GETconstructor.query_congestion_costs — Function

query_congestion_costs(domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = ""])

Constructs the HTTP request for the data about the congestion management costs over a certain border (article 13.1 C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/congestion-management/Data-view%20Costs.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one month!

Arguments

domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

! 100 document limit applies !

Parse functions

entsoeAPI.xmlParser.parse_expansion_and_dismantling — Function

parse_expansion_and_dismantling(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_expansion_and_dismantling(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", docStatus::String = ""]). Returns the data in a dataframe.

[estimated completion date, new NTC => [start, end, new NTC], transmission assets => [code, location, type]]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_expansion_and_dismantling(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", docStatus::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_redispatching — Function

parse_redispatching(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_redispatching(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = ""]). Returns the data in a dataframe.

[start, end, reason, impact => [start, end , impact], affected assets => [code, location, type]]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_redispatching(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_countertrading — Function

parse_countertrading(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_countertrading(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[start, end, reason, change in cross-border exchange => [start, end, change in cross-border exchange]]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_countertrading(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_congestion_costs — Function

parse_congestion_costs(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_congestion_costs(domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = ""]). Returns the data in a dataframe.

[start, end, costs]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_congestion_costs(domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented