Load functions

This file describes the usage of all the functions under the load tab in the ENTSOE-E transparancy platform.

User functions

entsoeAPI.actual_total_load — Function

actual_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for actual total load data (article 6.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Total%20Load%20-%20Day%20Ahead%20-%20Actual.html). Parses the received HTTP response and returns the data in a dataframe.

[time, load]

Minimum time interval in query response is one MTU period!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.day_ahead_total_load — Function

day_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the day ahaed total load forecast (article 6.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Total%20Load%20-%20Day%20Ahead%20-%20Actual.html). Parses the received HTTP response and returns the data in a dataframe.

[time, load]

Minimum time interval in query response is one day!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.week_ahead_total_load — Function

week_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the week ahaed total load forecast (article 6.1 C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Total%20Load%20Forecast%20-%20Week%20Ahead.html). Parses the received HTTP response and returns the data in a dictionary.

("min total load" => [time, load], "max total load" => [time, load])

Minimum time interval in query response is one week!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.month_ahead_total_load — Function

month_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the month ahaed total load forecast (article 6.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Total%20Load%20Forecast%20-%20Month%20Ahead.html). Parses the received HTTP response and returns the data in a dictionary.

("min total load" => [week, load], "max total load" => [week, load])

Minimum time interval in query response is one month!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.year_ahead_total_load — Function

year_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the year ahaed total load forecast (article 6.1 E: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Total%20Load%20Forecast%20-%20Year%20Ahead.html). Parses the received HTTP response and returns the data in a dictionary.

("min total load" => [week, load], "max total load" => [week, load])

Minimum time interval in query response is one year!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.year_ahead_margin — Function

year_ahead_margin(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the year ahead forecast margin (article 8.1: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Forecast%20Margin%20-%20Year%20Ahead.html). Parses the received HTTP response and returns the data in a dataframe.

[year, margin]

Minimum time interval in query response is one year!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

GET functions

entsoeAPI.GETconstructor.query_actual_total_load — Function

query_actual_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for actual total load data (article 6.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Total%20Load%20-%20Day%20Ahead%20-%20Actual.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one MTU period!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_day_ahead_total_load — Function

query_day_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the day ahaed total load forecast (article 6.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Total%20Load%20-%20Day%20Ahead%20-%20Actual.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one day!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_week_ahead_total_load — Function

query_week_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the week ahaed total load forecast (article 6.1 C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Total%20Load%20Forecast%20-%20Week%20Ahead.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one week!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_month_ahead_total_load — Function

query_month_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the month ahaed total load forecast (article 6.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Total%20Load%20Forecast%20-%20Month%20Ahead.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one month!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_year_ahead_total_load — Function

query_year_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the year ahaed total load forecast (article 6.1 E: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Total%20Load%20Forecast%20-%20Year%20Ahead.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one year!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_year_ahead_margin — Function

query_year_ahead_margin(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the year ahead forecast margin (article 8.1: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/load-domain/Data-view%20Forecast%20Margin%20-%20Year%20Ahead.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one year!

Arguments

outBiddingZone_Domain::Union{mappings.Area, String}: Area for which the load data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

Parse function

entsoeAPI.xmlParser.parse_actual_total_load — Function

parse_actual_total_load(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_actual_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[time, load]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_actual_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_day_ahead_total_load — Function

parse_day_ahead_total_load(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_day_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[time, load]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_day_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_week_ahead_total_load — Function

parse_week_ahead_total_load(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_week_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dictionary.

("min total load" => [time, load], "max total load" => [time, load])

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_week_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_month_ahead_total_load — Function

parse_month_ahead_total_load(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_month_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dictionary.

("min total load" => [week, load], "max total load" => [week, load])

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_month_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_year_ahead_total_load — Function

parse_year_ahead_total_load(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_year_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dictionary.

("min total load" => [week, load], "max total load" => [week, load])

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_month_ahead_total_load(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_year_ahead_margin — Function

parse_year_ahead_margin(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_year_ahead_margin(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in dataframe.

[year, margin]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_year_ahead_margin(outBiddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented