Generation functions

This file describes the usage of all the functions under the generation tab in the ENTSOE-E transparancy platform.

User functions

entsoeAPI.installed_generation_capacity_aggregated — Function

installed_generation_capacity_aggregated(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the aggregated installed generation capacity in a certain area (article 14.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Installed%20Capacity%20per%20Production%20Type.html). Parses the received HTTP response and returns the data in a dataframe.

[year, type, installed capacity]

Minimum time interval in query response is one year!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

! One year range limit applies !

entsoeAPI.installed_generation_capacity_per_unit — Function

installed_generation_capacity_per_unit(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the installed generation capacity per unit in a certain area (article 14.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Installed%20Capacity%20per%20Production%20Unit.html). Parses the received HTTP response and returns the data in a dataframe.

[production type, code, name, installed capacity at the beginning of the year, voltage connection level]

Minimum time interval in query response is one year!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

! One year range limit applies !

entsoeAPI.day_ahead_aggregated_generation — Function

day_ahead_aggregated_generation(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the day ahead forecast of aggregated generation in a certain area (article 14.1 C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Generation%20Forecast%20-%20Day%20Ahead.html). Parses the received HTTP response and returns the data in a dictionary.

("generation" => [start, end, scheduled generation], "consumption" => [start, end, scheduled consumption])

Minimum time interval in query response is one day!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.day_ahead_generation_forecasts_wind_solar — Function

day_ahead_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the day ahead forecast of wind and solar generation in a certain area (article 14.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Generation%20Forecasts%20-%20Day%20Ahead%20for%20Wind%20and%20Solar.html). Parses the received HTTP response and returns the data in a dictionary.

("solar" => [start, end, solar], "wind offshore" => [start, end, wind offshore], "wind onshore" => [start, end, wind onshore])

Minimum time interval in query response is one day!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

! One year range limit applies !

entsoeAPI.current_generation_forecasts_wind_solar — Function

current_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the current forecast of wind and solar generation in a certain area (article 14.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Generation%20Forecast%20-%20Day%20Ahead.html). Parses the received HTTP response and returns the data in a dictionary.

("solar" => [start, end, solar], "wind offshore" => [start, end, wind offshore], "wind onshore" => [start, end, wind onshore])

Minimum time interval in query response is one day!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

! One year range limit applies !

entsoeAPI.intraday_generation_forecasts_wind_solar — Function

intraday_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the intraday forecast of wind and solar generation in a certain area (article 14.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Generation%20Forecast%20-%20Day%20Ahead.html). Parses the received HTTP response and returns the data in a dictionary.

("solar" => [start, end, solar], "wind offshore" => [start, end, wind offshore], "wind onshore" => [start, end, wind onshore])

Minimum time interval in query response is one MTU period!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

! One year range limit applies !

entsoeAPI.actual_generation_per_generation_unit — Function

actual_generation_per_generation_unit(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = "", registeredResource::String = ""])

Constructs the HTTP request for the data about the actual generation per generation unit in a certain area (article 16.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Actual%20Generation%20per%20Generation%20Unit.html). Parses the received HTTP response and returns the data in a dictionary.

NOT IMPLEMENTED YET

Minimum time interval in query response is one MTU period!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

registeredResource = "": The unique identification of a resource

! One day range limit applies !

entsoeAPI.aggregated_generation_per_type — Function

aggregated_generation_per_type(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the actual aggregated generation per plant type in a certain area (article 16.1 B&C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Actual%20Generation%20per%20Production%20Unit.html). Parses the received HTTP response and returns the data in a dictionary.

("type" => [start, end, aggregated generation])

Minimum time interval in query response is one MTU period!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

! One year range limit applies !

entsoeAPI.aggregated_filling_rate — Function

aggregated_filling_rate(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the aggregated filling rate of water reservoirs and hydro storage plants in a certain area (article 16.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Water%20Reservoirs%20and%20Hydro%20Storage%20Plants.html). Parses the received HTTP response and returns the data in a dictionary.

[year, week, stored energy value]

Minimum time inteval in query response is one week!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

GET functions

entsoeAPI.GETconstructor.query_installed_generation_capacity_aggregated — Function

query_installed_generation_capacity_aggregated(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the aggregated installed generation capacity in a certain area (article 14.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Installed%20Capacity%20per%20Production%20Type.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one year!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset.

! One year range limit applies !

entsoeAPI.GETconstructor.query_installed_generation_capacity_per_unit — Function

query_installed_generation_capacity_per_unit(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the installed generation capacity per unit in a certain area (article 14.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Installed%20Capacity%20per%20Production%20Unit.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one year!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

! One year range limit applies !

entsoeAPI.GETconstructor.query_day_ahead_aggregated_generation — Function

query_day_ahead_aggregated_generation(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the day ahead forecast of aggregated generation in a certain area (article 14.1 C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Generation%20Forecast%20-%20Day%20Ahead.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one day!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_day_ahead_generation_forecasts_wind_solar — Function

query_day_ahead_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the day ahead forecast of wind and solar generation in a certain area (article 14.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Generation%20Forecasts%20-%20Day%20Ahead%20for%20Wind%20and%20Solar.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one day!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

! One year range limit applies !

entsoeAPI.GETconstructor.query_current_generation_forecasts_wind_solar — Function

query_current_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the current forecast of wind and solar generation in a certain area (article 14.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Generation%20Forecast%20-%20Day%20Ahead.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one day!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

! One year range limit applies !

entsoeAPI.GETconstructor.query_intraday_generation_forecasts_wind_solar — Function

query_intraday_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the intraday forecast of wind and solar generation in a certain area (article 14.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Generation%20Forecast%20-%20Day%20Ahead.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one MTU period!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

! One year range limit applies !

entsoeAPI.GETconstructor.query_actual_generation_per_generation_unit — Function

query_actual_generation_per_generation_unit(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = "", registeredResource::String = ""])

Constructs the HTTP request for the data about the actual generation per generation unit in a certain area (article 16.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Actual%20Generation%20per%20Generation%20Unit.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one MTU period!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

registeredResource = "": The unique identification of a resource

! One day range limit applies !

entsoeAPI.GETconstructor.query_aggregated_generation_per_type — Function

query_aggregated_generation_per_type(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

Constructs the HTTP request for the data about the actual aggregated generation per plant type in a certain area (article 16.1 B&C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Actual%20Generation%20per%20Production%20Unit.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one MTU period!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

psrType::String = "": The coded type of a power system resource. The classification for the asset. If not used all resources are included.

! One year range limit applies !

entsoeAPI.GETconstructor.query_aggregated_filling_rate — Function

query_aggregated_filling_rate(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data about the aggregated filling rate of water reservoirs and hydro storage plants in a certain area (article 16.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/generation/Data-view%20Water%20Reservoirs%20and%20Hydro%20Storage%20Plants.html). Returns the received HTTP response, together with the used timezone. Minimum time inteval in query response is one week!

Arguments

in_Domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

Parse functions

entsoeAPI.xmlParser.parse_installed_generation_capacity_aggregated — Function

parse_installed_generation_capacity_aggregated(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_installed_generation_capacity_aggregated(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""]). Returns the data in a dataframe.

[year, type, installed capacity]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_installed_generation_capacity_aggregated(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_installed_generation_capacity_per_unit — Function

parse_installed_generation_capacity_per_unit(xml::Vector{UInt8})

Parses xml file generated by the function GETconstructor.query_installed_generation_capacity_per_unit(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""]). Returns the data in a dataframe.

[production type, code, name, installed capacity at the beginnen of the year, voltage connection level]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_installed_generation_capacity_per_unit(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_day_ahead_aggregated_generation — Function

parse_day_ahead_aggregated_generation(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_day_ahead_aggregated_generation(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dictionary.

("generation" => [start, end, scheduled generation], "consumption" => [start, end, scheduled consumption])

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_day_ahead_aggregated_generation(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_day_ahead_generation_forecasts_wind_solar — Function

parse_day_ahead_generation_forecasts_wind_solar(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_day_ahead_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime, psrType::String = ""). Returns the data in a dictionary.

("solar" => [start, end, solar], "wind offshore" => [start, end, wind offshore], "wind onshore" => [start, end, wind onshore])

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_day_ahead_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime, psrType::String = "")

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_current_generation_forecasts_wind_solar — Function

parse_current_generation_forecasts_wind_solar(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_current_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""]). Returns the data in a dictionary.

("solar" => [start, end, solar], "wind offshore" => [start, end, wind offshore], "wind onshore" => [start, end, wind onshore])

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_current_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_intraday_generation_forecasts_wind_solar — Function

parse_intraday_generation_forecasts_wind_solar(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_intraday_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""]). Returns the data in a dictionary.

("solar" => [start, end, solar], "wind offshore" => [start, end, wind offshore], "wind onshore" => [start, end, wind onshore])

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_intraday_generation_forecasts_wind_solar(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_actual_generation_per_generation_unit — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_aggregated_generation_per_type — Function

parse_aggregated_generation_per_type(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_aggregated_generation_per_type(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""]). Returns the data in a dictionary.

("type" => [start, end, aggregated generation])

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_aggregated_generation_per_type(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, psrType::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_aggregated_filling_rate — Function

parse_aggregated_filling_rate(xml::Vector{UInt8}, tz::TimeZone)

Parses xml file generated by the function GETconstructor.query_aggregated_filling_rate(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[year, week, stored energy value]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_aggregated_filling_rate(in_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented