Outages functions

This file describes the usage of all the functions under the outages tab in the ENTSOE-E transparancy platform.

User functions

entsoeAPI.unavailability_consumption_units — Function

unavailability_consumption_units(biddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = ""])

Constructs the HTTP request for the data about the unavailability of consumption units in a certain area (article 7.1 A&B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/outage-domain/Data-view%20Aggregated%20Unavailability%20of%20Consumption%20Units.html). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Minimum time interval in query response is one MTU period!

Arguments

biddingZone_Domain::Union{mappings.Area, String}: Area for which the outages data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

! One year range limit applies !

entsoeAPI.unavailability_generation_units — Function

unavailability_generation_units(biddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", docStatus::String = "", periodStartUpdate::DateTime = DateTime(0), periodEndUpdate::DateTime = DateTime(0), registeredResource::String = "", mRID::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the unavailability of generation units in a certain area (article 15.1 A&B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/outage-domain/Data-view%20Unavailability%20of%20Production%20and%20Generation%20Units.html). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Minimum time interval in query response depends on duration of matching outages!

Arguments

biddingZone_Domain::Union{mappings.Area, String}: Area for which the outages data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

docStatus::String = "": Identification of the condition or position of the document with regard to its standing

periodStartUpdate::DateTime = DatetTime(0):

periodEndUpdate::DateTime = DateTime(0):

registeredResource::String = "": The unique identification of a resource

mrRID::String = "": Unique identification of the document being exchanged within a business process flow

offset::Int = 0: allows downloading more than 200 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+200

! One year range limit applies ! ! 200 documents limit applies !

entsoeAPI.unavailability_production_units — Function

unavailability_production_units(biddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", docStatus::String = "", periodStartUpdate::DateTime = DateTime(0), periodEndUpdate::DateTime = DateTime(0), registeredResource::String = "", mRID::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the unavailability of production units in a certain area (article 15.1 C&D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/outage-domain/Data-view%20Unavailability%20of%20Production%20and%20Generation%20Units.html). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Minimum time interval in query response depends on duration of matching outages!

Arguments

biddingZone_Domain::Union{mappings.Area, String}: Area for which the outages data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

docStatus::String = "": Identification of the condition or position of the document with regard to its standing

periodStartUpdate::DateTime = DatetTime(0):

periodEndUpdate::DateTime = DateTime(0):

registeredResource::String = "": The unique identification of a resource

mrRID::String = "": Unique identification of the document being exchanged within a business process flow

offset::Int = 0: allows downloading more than 200 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+200

! One year range limit applies ! ! 200 documents limit applies !

entsoeAPI.unavailability_offshore_grid — Function

unavailability_offshore_grid(biddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, docStatus::String = "", periodStartUpdate::DateTime = DateTime(0), periodEndUpdate::DateTime = DateTime(0), mRID::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the unavailability of the offshore grid infrastructure in a certain area (article 10.1 C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/outage-domain/Data-view%20Unavailability%20of%20off-shore%20grid.html). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Minimum time interval in query response depends on duration of matching outages!

Arguments

biddingZone_Domain::Union{mappings.Area, String}: Area for which the outages data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

docStatus::String = "": Identification of the condition or position of the document with regard to its standing

periodStartUpdate::DateTime = DatetTime(0):

periodEndUpdate::DateTime = DateTime(0):

registeredResource::String = "": The unique identification of a resource

mrRID::String = "": Unique identification of the document being exchanged within a business process flow

offset::Int = 0: allows downloading more than 200 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+200

! One year range limit applies ! ! 200 documents limit applies !

entsoeAPI.unavailability_transmission_infrastructure — Function

unavailability_transmission_infrastructure(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", docStatus::String = "", periodStartUpdate::DateTime = DateTime(0), periodEndUpdate::DateTime = DateTime(0), mRID::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the unavailability of the transmission infrastructure in a certain area (article 10.1 A&B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/outage-domain/Data-view%20Unavailability%20in%20Transmission%20Grid.html). Parses the received HTTP response and returns the data in a dataframe.

NOT YET IMPLEMENTED

Minimum time interval in query response depends on duration of matching outages!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

docStatus::String = "": Identification of the condition or position of the document with regard to its standing

periodStartUpdate::DateTime = DatetTime(0):

periodEndUpdate::DateTime = DateTime(0):

registeredResource::String = "": The unique identification of a resource

mrRID::String = "": Unique identification of the document being exchanged within a business process flow

offset::Int = 0: allows downloading more than 200 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+200

! One year range limit applies ! ! 200 documents limit applies !

entsoeAPI.fallBacks — Function

STILL NEEDS TO BE IMPLEMENTED

GET functions

entsoeAPI.GETconstructor.query_unavailability_consumption_units — Function

query_unavailability_consumption_units(biddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = ""])

Constructs the HTTP request for the data about the unavailability of consumption units in a certain area (article 7.1 A&B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/outage-domain/Data-view%20Aggregated%20Unavailability%20of%20Consumption%20Units.html). Returns the received HTTP response. Minimum time interval in query response is one MTU period!

Arguments

biddingZone_Domain::Union{mappings.Area, String}: Area for which the outages data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

! One year range limit applies !

entsoeAPI.GETconstructor.query_unavailability_generation_units — Function

query_unavailability_generation_units(biddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", docStatus::String = "", periodStartUpdate::DateTime = DateTime(0), periodEndUpdate::DateTime = DateTime(0), registeredResource::String = "", mRID::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the unavailability of generation units in a certain area (article 15.1 A&B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/outage-domain/Data-view%20Unavailability%20of%20Production%20and%20Generation%20Units.html). Returns the received HTTP response. Minimum time interval in query response depends on duration of matching outages!

Arguments

biddingZone_Domain::Union{mappings.Area, String}: Area for which the outages data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

docStatus::String = "": Identification of the condition or position of the document with regard to its standing

periodStartUpdate::DateTime = DatetTime(0):

periodEndUpdate::DateTime = DateTime(0):

registeredResource::String = "": The unique identification of a resource

mrRID::String = "": Unique identification of the document being exchanged within a business process flow

offset::Int = 0: allows downloading more than 200 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+200

! One year range limit applies ! ! 200 documents limit applies !

entsoeAPI.GETconstructor.query_unavailability_production_units — Function

query_unavailability_production_units(biddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", docStatus::String = "", periodStartUpdate::DateTime = DateTime(0), periodEndUpdate::DateTime = DateTime(0), registeredResource::String = "", mRID::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the unavailability of production units in a certain area (article 15.1 C&D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/outage-domain/Data-view%20Unavailability%20of%20Production%20and%20Generation%20Units.html). Returns the received HTTP response. Minimum time interval in query response depends on duration of matching outages!

Arguments

biddingZone_Domain::Union{mappings.Area, String}: Area for which the outages data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

docStatus::String = "": Identification of the condition or position of the document with regard to its standing

periodStartUpdate::DateTime = DatetTime(0):

periodEndUpdate::DateTime = DateTime(0):

registeredResource::String = "": The unique identification of a resource

mrRID::String = "": Unique identification of the document being exchanged within a business process flow

offset::Int = 0: allows downloading more than 200 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+200

! One year range limit applies ! ! 200 documents limit applies !

entsoeAPI.GETconstructor.query_unavailability_offshore_grid — Function

query_unavailability_offshore_grid(biddingZone_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, docStatus::String = "", periodStartUpdate::DateTime = DateTime(0), periodEndUpdate::DateTime = DateTime(0), mRID::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the unavailability of the offshore grid infrastructure in a certain area (article 10.1 C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/outage-domain/Data-view%20Unavailability%20of%20off-shore%20grid.html). Returns the received HTTP response. Minimum time interval in query response depends on duration of matching outages!

Arguments

biddingZone_Domain::Union{mappings.Area, String}: Area for which the outages data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

docStatus::String = "": Identification of the condition or position of the document with regard to its standing

periodStartUpdate::DateTime = DatetTime(0):

periodEndUpdate::DateTime = DateTime(0):

registeredResource::String = "": The unique identification of a resource

mrRID::String = "": Unique identification of the document being exchanged within a business process flow

offset::Int = 0: allows downloading more than 200 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+200

! One year range limit applies ! ! 200 documents limit applies !

entsoeAPI.GETconstructor.query_unavailability_transmission_infrastructure — Function

query_unavailability_transmission_infrastructure(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, businessType::String = "", docStatus::String = "", periodStartUpdate::DateTime = DateTime(0), periodEndUpdate::DateTime = DateTime(0), mRID::String = "", offset::Int = 0])

Constructs the HTTP request for the data about the unavailability of the transmission infrastructure in a certain area (article 10.1 A&B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/outage-domain/Data-view%20Unavailability%20in%20Transmission%20Grid.html). Returns the received HTTP response. Minimum time interval in query response depends on duration of matching outages!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

businessType::String = "": The identification of the nature of the data

docStatus::String = "": Identification of the condition or position of the document with regard to its standing

periodStartUpdate::DateTime = DatetTime(0):

periodEndUpdate::DateTime = DateTime(0):

registeredResource::String = "": The unique identification of a resource

mrRID::String = "": Unique identification of the document being exchanged within a business process flow

offset::Int = 0: allows downloading more than 200 documents. The offset ∈ [0,4800] so that paging is restricted to query for 4900 documents max., offset=n returns files in sequence between n+1 and n+200

! One year range limit applies ! ! 200 documents limit applies !

entsoeAPI.GETconstructor.query_fallBacks — Function

STILL NEEDS TO BE IMPLEMENTED

Parse functions

entsoeAPI.xmlParser.parse_unavailability_consumption_units — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_unavailability_generation_units — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_unavailability_production_units — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_unavailability_offshore_grid — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_unavailability_transmission_infrastructure — Function

STILL NEEDS TO BE IMPLEMENTED

entsoeAPI.xmlParser.parse_fallBacks — Function

STILL NEEDS TO BE IMPLEMENTED