Transmission functions

This file describes the usage of all the functions under the load tab in the ENTSOE-E transparancy platform.

User functions

entsoeAPI.forecasted_capacity — Function

forecasted_capacity(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the forecasted capacity over a certain border (article 11.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Forecasted%20Transfer%20Capacities%20-%20Day%20Ahead.html). Parses the received HTTP response and returns the data in a dataframe.

[time, capacity]

Minimum time interval in query response ranges from day to year, depending on selected Contract_MarketAgreementType!

Arguments

contract_MarketAgreementType::String:The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.offered_capacity — Function

offered_capacity(auctionType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", update_DateAndOrTime::DateTime = DateTime(0), classificationSequence_AttributeInstanceComponentPosition::String = ""])

Constructs the HTTP request for the data of the offered capacity over a certain border (article 11.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission/Data-view%20Explicit%20Allocations%20-%20Intraday.html). Parses the received HTTP response and returns the data in a dataframe.

[time, capacity]

Minimum time interval in query response ranges from day to year, depending on selected Contract_MarketAgreementType!

Arguments

auctionType::String: The kind of the auction (e.g. implicit, explicit ...)

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

auctionCategory::String = "": The product category of an auction

update_DateAndOrTime::DateTime = "": The date and time of the update of the document

classificationSequence_AttributeInstanceComponentPosition::String = "": A sequential value representing a relative sequence number. A classification sequence is only provided in the case where there are several auctions in the same category and contract type.

! 100 document limit applies !

entsoeAPI.flowbased — Function

flowbased(processType::String, domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the flow-based parameters of a certain area (article 11.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Day%20Ahead%20Flow%20Based%20Allocations.html). Parses the received HTTP response and returns the data in a dictionary.

("time" => [cb/co, RAM, domain, ..., domain])

Minimum time interval in query response is one day for day-ahead allocations!

Arguments

processType::String: The kind of the auction (e.g. implicit, explicit ...)

domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! 100 document limit applies !

entsoeAPI.intraday_transfer_limits — Function

intraday_transfer_limits(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the intraday transfer limits over a certain border (article 11.3: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Cross%20Border%20Capacity%20of%20DC%20Links%20-%20Intraday%20Transfer%20Limits.html). Parses the received HTTP response and returns the data in a dataframe.

[time, capacity]

Minimum time interval in query response ranges from part of day up to one day!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.explicit_allocation_information_capacity — Function

explicit_allocation_information_capacity(businessType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", classificationSequence_AttributeInstanceComponentPosition::String = ""])

Constructs the HTTP request for the data of the capacity explicitly allocated to the market over a certain border and its revenue (article 12.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission/Data-view%20Explicit%20Allocations%20-%20Day%20ahead.html). Parses the received HTTP response and returns the data in a dataframe.

[time, capacity [, price]]

Minimum time interval in query response ranges from part of day to year, depending on selected Contract_MarketAgreementType!

Arguments

businessType::String: The identification of the nature of the data

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

auctionCategory::String = "": The product category of an auction

classificationSequence_AttributeInstanceComponentPosition::String = "": A sequential value representing a relative sequence number. A classification sequence is only provided in the case where there are several auctions in the same category and contract type.

! 100 document limit applies !

entsoeAPI.explicit_allocation_information_revenue — Function

explicit_allocation_information_revenue(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the revenue of the capacity explicitly allocated to the market over a certain border (article 12.1 A: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Explicit%20Allocations%20Revenue.html). Parses the received HTTP response and returns the data in a dataframe.

[time, price]

Minimum time interval in query response ranges from part of day to year, depending on selected Contract_MarketAgreementType!

Arguments

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! 100 document limit applies !

entsoeAPI.total_capacity_nominated — Function

total_capacity_nominated(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the total nominated capacity over a certain border (article 12.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Total%20Nominated%20Capacity.html). Parses the received HTTP response and returns the data in a dataframe.

[time, capacity]

Minimum time interval in query response is one day!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.total_capacity_already_allocated — Function

total_capacity_already_allocated(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = ""])

Constructs the HTTP request for the data of the total capacity already allocated over a certain border (article 12.1 C: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission/Data-view%20Explicit%20Allocations%20-%20AAC.html). Parses the received HTTP response and returns the data in a dataframe.

[time, capacity]

Minimum time interval in query response ranges from part of day to year, depending on selected Contract_MarketAgreement.Type!

Arguments

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

auctionCategory::String = "": The product category of an auction

! 100 documents limit applies !

entsoeAPI.day_ahead_prices — Function

day_ahead_prices(domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the day ahead prices in a certain area (article 12.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Day-ahead%20prices.html). Parses the received HTTP response and returns the data in a dataframe.

[time, price]

Minimum time interval in query response is one day!

Arguments

domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.implicit_auction_net_positions — Function

implicit_auction_net_positions(businessType::String, contract_MarketAgreementType::String, domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the net positions and congestion income of implictly allocated capacity over a certain border (article 12.1 E: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Intraday%20Flow%20Based%20Implicit%20Allocations%20-%20Congestion%20Income.html). Parses the received HTTP response and returns the data in a dataframe.

[time, import/export, capacity]

Minimum time interval in query response is one day!

Arguments

businessType::String: The identification of the nature of the data.

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies ! ! 100 documents limit applies !

entsoeAPI.implicit_auction_congestion_income — Function

implicit_auction_congestion_income(businessType::String, contract_MarketAgreementType::String, domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the net positions and congestion income of implictly allocated capacity over a certain border (article 12.1 E: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Intraday%20Flow%20Based%20Implicit%20Allocations%20-%20Congestion%20Income.html). Parses the received HTTP response and returns the data in a dataframe.

[time, price]

Minimum time interval in query response is one day!

Arguments

businessType::String: The identification of the nature of the data.

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies ! ! 100 documents limit applies !

entsoeAPI.total_commercial_schedules — Function

total_commercial_schedules(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the total or day ahead commercial schedules over a certain border (article 12.1 F: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Scheduled%20Commercial%20Exchanges%20-%20Day%20Ahead.html). Parses the received HTTP response and returns the data in a dataframe.

[time, capacity]

Minimum time interval in query response is one day!

Arguments

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

Missing docstring.

Missing docstring for entsoeAPI.physical_flows. Check Documenter's build log for details.

entsoeAPI.capacity_allocated_outside_EU — Function

capacity_allocated_outside_EU(auctionType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", classificationSequence_AttributeInstanceComponentPosition::String = ""])

Constructs the HTTP request for the data of the capacity allocated outside the EU over a certain border (article 12.1 H: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission/Data-view%20Transfer%20Capacities%20Allocated%20with%20Third%20Countries.html). Parses the received HTTP response and returns the data in a dataframe.

[time, capacity]

Minimum time interval in query response ranges from part of day to year, depending on selected Contract_MarketAgreementType!

Arguments

auctionType::String: The kind of the auction (e.g. implicit, explicit ...)

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

auctionCategory::String = "": The product category of an auction

classificationSequence_AttributeInstanceComponentPosition::String = "": A sequential value representing a relative sequence number. A classification sequence is only provided in the case where there are several auctions in the same category and contract type.

! 100 document limit applies !

GET functions

entsoeAPI.GETconstructor.query_forecasted_capacity — Function

query_forecasted_capacity(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the forecasted capacity over a certain border (article 11.1 A: https://transparency.entsoe.eu/content/staticcontent/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Forecasted%20Transfer%20Capacities%20-%20Day%20Ahead.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response ranges from day to year, depending on selected ContractMarketAgreementType!

Arguments

contract_MarketAgreementType::String:The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_offered_capacity — Function

query_offered_capacity(auctionType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", update_DateAndOrTime::DateTime = DateTime(0), classificationSequence_AttributeInstanceComponentPosition::String = ""])

Constructs the HTTP request for the data of the offered capacity over a certain border (article 11.1 A: https://transparency.entsoe.eu/content/staticcontent/Static%20content/knowledge%20base/data-views/transmission/Data-view%20Explicit%20Allocations%20-%20Intraday.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response ranges from day to year, depending on selected ContractMarketAgreementType!

Arguments

auctionType::String: The kind of the auction (e.g. implicit, explicit ...)

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

auctionCategory::String = "": The product category of an auction

update_DateAndOrTime::DateTime = "": The date and time of the update of the document

classificationSequence_AttributeInstanceComponentPosition::String = "": A sequential value representing a relative sequence number. A classification sequence is only provided in the case where there are several auctions in the same category and contract type.

! 100 document limit applies !

entsoeAPI.GETconstructor.query_flowbased — Function

query_flowbased(processType::String, domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the flow-based parameters of a certain area (article 11.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Day%20Ahead%20Flow%20Based%20Allocations.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one day for day-ahead allocations!

Arguments

processType::String: The kind of the auction (e.g. implicit, explicit ...)

domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! 100 document limit applies !

entsoeAPI.GETconstructor.query_intraday_transfer_limits — Function

query_intraday_transfer_limits(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the intraday transfer limits over a certain border (article 11.3: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Cross%20Border%20Capacity%20of%20DC%20Links%20-%20Intraday%20Transfer%20Limits.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response ranges from part of day up to one day!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_explicit_allocation_information_capacity — Function

query_explicit_allocation_information_capacity(businessType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", classificationSequence_AttributeInstanceComponentPosition::String = ""])

Constructs the HTTP request for the data of the capacity explicitly allocated to the market over a certain border and its revenue (article 12.1 A: https://transparency.entsoe.eu/content/staticcontent/Static%20content/knowledge%20base/data-views/transmission/Data-view%20Explicit%20Allocations%20-%20Day%20ahead.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response ranges from part of day to year, depending on selected ContractMarketAgreementType!

Arguments

businessType::String: The identification of the nature of the data

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

auctionCategory::String = "": The product category of an auction

classificationSequence_AttributeInstanceComponentPosition::String = "": A sequential value representing a relative sequence number. A classification sequence is only provided in the case where there are several auctions in the same category and contract type.

! 100 document limit applies !

entsoeAPI.GETconstructor.query_explicit_allocation_information_revenue — Function

query_explicit_allocation_information_revenue(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the revenue of the capacity explicitly allocated to the market over a certain border (article 12.1 A: https://transparency.entsoe.eu/content/staticcontent/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Explicit%20Allocations%20Revenue.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response ranges from part of day to year, depending on selected ContractMarketAgreementType!

Arguments

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! 100 document limit applies !

entsoeAPI.GETconstructor.query_total_capacity_nominated — Function

query_total_capacity_nominated(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the total nominated capacity over a certain border (article 12.1 B: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Total%20Nominated%20Capacity.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one day!

Arguments

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

entsoeAPI.GETconstructor.query_total_capacity_already_allocated — Function

query_total_capacity_already_allocated(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = ""])

Constructs the HTTP request for the data of the total capacity already allocated over a certain border (article 12.1 C: https://transparency.entsoe.eu/content/staticcontent/Static%20content/knowledge%20base/data-views/transmission/Data-view%20Explicit%20Allocations%20-%20AAC.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response ranges from part of day to year, depending on selected ContractMarketAgreement.Type!

Arguments

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

auctionCategory::String = "": The product category of an auction

! 100 documents limit applies !

entsoeAPI.GETconstructor.query_day_ahead_prices — Function

query_day_ahead_prices(domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the day ahead prices in a certain area (article 12.1 D: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Day-ahead%20prices.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one day!

Arguments

domain::Union{mappings.Area, String}: The area for which the data is needed, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

Missing docstring.

Missing docstring for entsoeAPI.GETconstructor.query_implicit_auction_net_positions. Check Documenter's build log for details.

Missing docstring.

Missing docstring for entsoeAPI.GETconstructor.query_implicit_auction_congestion_income. Check Documenter's build log for details.

entsoeAPI.GETconstructor.query_total_commercial_schedules — Function

query_total_commercial_schedules(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

Constructs the HTTP request for the data of the total or day ahead commercial schedules over a certain border (article 12.1 F: https://transparency.entsoe.eu/content/static_content/Static%20content/knowledge%20base/data-views/transmission-domain/Data-view%20Scheduled%20Commercial%20Exchanges%20-%20Day%20Ahead.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response is one day!

Arguments

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

! One year range limit applies !

Missing docstring.

Missing docstring for entsoeAPI.GETconstructor.query_physical_flows. Check Documenter's build log for details.

entsoeAPI.GETconstructor.query_capacity_allocated_outside_EU — Function

query_capacity_allocated_outside_EU(auctionType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", classificationSequence_AttributeInstanceComponentPosition::String = ""])

Constructs the HTTP request for the data of the capacity allocated outside the EU over a certain border (article 12.1 H: https://transparency.entsoe.eu/content/staticcontent/Static%20content/knowledge%20base/data-views/transmission/Data-view%20Transfer%20Capacities%20Allocated%20with%20Third%20Countries.html). Returns the received HTTP response, together with the used timezone. Minimum time interval in query response ranges from part of day to year, depending on selected ContractMarketAgreementType!

Arguments

auctionType::String: The kind of the auction (e.g. implicit, explicit ...)

contract_MarketAgreementType::String: The specification of the kind of the agreement, e.g. long term, daily contract. Used to distinguish between day ahead, week ahead, month ahead and year ahead forecasts.

in_Domain::Union{mappings.Area, String}: The area where energy is going, can be represented as an Area object or a string with country code or direct code

out_Domain::Union{mappings.Area, String}: The area where energy is coming from, can be represented as an Area object or a string with country code or direct code

periodStart::DateTime: Start date and time of the needed data

periodEnd::DateTime: End date and time of the needed data

auctionCategory::String = "": The product category of an auction

classificationSequence_AttributeInstanceComponentPosition::String = "": A sequential value representing a relative sequence number. A classification sequence is only provided in the case where there are several auctions in the same category and contract type.

! 100 document limit applies !

Parse functions

entsoeAPI.xmlParser.parse_forecasted_capacity — Function

parse_forecasted_capacity(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_forecasted_capacity(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[time, capacity]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_forecasted_capacity(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_offered_capacity — Function

parse_offered_capacity(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_offered_capacity(auctionType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", update_DateAndOrTime::DateTime = DateTime(0), classificationSequence_AttributeInstanceComponentPosition::String = ""]). Returns the data in a dataframe.

[time, capacity]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_offered_capacity(auctionType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", update_DateAndOrTime::DateTime = DateTime(0), classificationSequence_AttributeInstanceComponentPosition::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_flowbased — Function

parse_flowbased(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_flowbased(processType::String, domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in dictionary.

("time" => [cb/co, RAM, domain, ..., domain])

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_flowbased(processType::String, domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_intraday_transfer_limits — Function

parse_intraday_transfer_limits(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_intraday_transfer_limits(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[time, capacity]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_intraday_transfer_limits(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_explicit_allocation_information_capacity — Function

parse_explicit_allocation_information_capacity(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_explicit_allocation_information_capacity(businessType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", classificationSequence_AttributeInstanceComponentPosition::String = ""]). Returns the data in a dataframe.

[time, capacity [, price]]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_explicit_allocation_information_capacity(businessType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", classificationSequence_AttributeInstanceComponentPosition::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_explicit_allocation_information_revenue — Function

parse_explicit_allocation_information_revenue(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_explicit_allocation_information_revenue(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime) Returns the data in a dataframe.

[time, price]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_explicit_allocation_information_revenue(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_total_capacity_nominated — Function

parse_total_capacity_nominated(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_total_capacity_nominated(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[time, capacity]

Arguments

- `xml::Vector{UInt8}`: xml data in the format as returned by the function `GETconstructor.query_total_capacity_nominated(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)`
- `tz::TimeZone`: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_total_capacity_already_allocated — Function

parse_total_capacity_already_allocated(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_total_capacity_already_allocated(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = ""]). Returns the data in a dataframe.

[time, capacity]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_total_capacity_already_allocated(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_day_ahead_prices — Function

parse_day_ahead_prices(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_day_ahead_prices(domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[time, capacity]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_day_ahead_prices(domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_implicit_auction_net_positions — Function

parse_implicit_auction_net_positions(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_implicit_auction_net_positions_and_congestion_income(businessType::String, contract_MarketAgreementType::String, domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime), in case the option 'net positions' is chosen. Returns the data in a dataframe.

[time, import/export, capacity]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_implicit_auction_net_positions_and_congestion_income(businessType::String, contract_MarketAgreementType::String, domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_implicit_auction_congestion_income — Function

parse_implicit_auction_congestion_income(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_implicit_auction_net_positions_and_congestion_income(businessType::String, contract_MarketAgreementType::String, domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime), in case the option 'congestion income' is chosen. Returns the data in a dataframe.

[time, price]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_implicit_auction_net_positions_and_congestion_income(businessType::String, contract_MarketAgreementType::String, domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_total_commercial_schedules — Function

parse_total_commercial_schedules(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_total_commercial_schedules(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[time, capacity]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_total_commercial_schedules(contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_physical_flows — Function

parse_physical_flows(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_phyiscal_flows(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime). Returns the data in a dataframe.

[time, capacity]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_phyiscal_flows(in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime)

tz::TimeZone: Timezone in which the dates and times have to be represented

entsoeAPI.xmlParser.parse_capacity_allocated_outside_EU — Function

parse_capacity_allocated_outside_EU(xml::Vector{UInt8}, tz::TimeZone)

Parses the xml file generated by the function GETconstructor.query_capacity_allocated_outside_EU(auctionType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", classificationSequence_AttributeInstanceComponentPosition::String = ""]). Returns the data in a dataframe.

[time, capacity]

Arguments

xml::Vector{UInt8}: xml data in the format as returned by the function GETconstructor.query_capacity_allocated_outside_EU(auctionType::String, contract_MarketAgreementType::String, in_Domain::Union{mappings.Area, String}, out_Domain::Union{mappings.Area, String}, periodStart::DateTime, periodEnd::DateTime[, auctionCategory::String = "", classificationSequence_AttributeInstanceComponentPosition::String = ""])

tz::TimeZone: Timezone in which the dates and times have to be represented